Change IsLambda* bounds from FnMut to Fn
The 'static bound basically means that they can't capture references
anyhow, so restricting the lambdas to be Fn instead of FnMut doesn't
really impose any increased restrictions.

It does allow these functions to then be shared when wrapped by just
an Arc instead of needing an Arc<Mutex>, which allows us to get rid of
a few locks.